move cleanDir functions out of extension.ts
