+++ __tests__/__integration__/extension.integration.test.ts
@@ -63,7 +63,7 @@ describe("Extension Integration Tests", () => {
     beforeEach(async function() {
         this.timeout(TIMEOUT);
         sandbox = sinon.createSandbox();
-        await extension.cleanTempDir();
+        await utils.cleanTempDir();
     });
 
     afterEach(async function() {
@@ -683,9 +683,9 @@ describe("Extension Integration Tests", () => {
         const providedPathTwo = path.join(__dirname, "..", "..", "..", "test-folder-two");
 
         // remove directories in case of previously failed tests
-        extension.cleanDir(testingPath);
-        extension.cleanDir(providedPathOne);
-        extension.cleanDir(providedPathTwo);
+        utils.cleanDir(testingPath);
+        utils.cleanDir(providedPathOne);
+        utils.cleanDir(providedPathTwo);
 
         it("should assign the temp folder based on preference", async () => {
             // create target folder
@@ -697,7 +697,7 @@ describe("Extension Integration Tests", () => {
             expect(ZOWETEMPFOLDER).to.equal(path.join(testingPath, "temp"));
 
             // Remove directory for subsequent tests
-            extension.cleanDir(testingPath);
+            utils.cleanDir(testingPath);
         }).timeout(TIMEOUT);
 
         it("should update temp folder on preference change", async () => {
@@ -716,8 +716,8 @@ describe("Extension Integration Tests", () => {
             expect(ZOWETEMPFOLDER).to.equal(path.join(providedPathTwo, "temp"));
 
             // Remove directory for subsequent tests
-            extension.cleanDir(providedPathOne);
-            extension.cleanDir(providedPathTwo);
+            utils.cleanDir(providedPathOne);
+            utils.cleanDir(providedPathTwo);
         }).timeout(TIMEOUT);
 
         it("should assign default temp folder, if preference is empty", async () => {
